Push TIM, clean code
