+++ main.c
@@ -4,10 +4,12 @@
 #include <string.h>
 #include <math.h>
 #include <ctype.h>
+#include "Structures.h"
+#include "affiche_syntaxe.h"
 #include "syntaxe.h"
 #include "priorite.h"
 #include "test_syntaxe.h"
-#include "Structures.h"
+
 
 int main(int argc, char *argv[]){
     tester_syntaxe();
@@ -15,13 +17,3 @@ int main(int argc, char *argv[]){
     return 0;
 }
 
-// TODO !!
-// 1
-// PENSER A ALLOC + FREE 
-// `-> fix erreur segmentation
-
-// 2
-// modif les tests pour mettre a jour les "erreurs" parce qu'il faut changer le type d'erreur
-
-// 3
-// ajouter des tests
